+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.16.0
+-----
+ * Improve display of ordered lists with large bullet numbers
+ * Fix bug where links with text that had a mix of Latin and non-Latin characters were getting split.
+ 
 1.15.0
 -----
  * Allow to use headers fonts without bold effect applied